refactor: correct preedit view behavior and polish its appearance
- Fix the view wouldn't disappear after changing theme or color on site
- Allow to custom its background color
- Add top left corner radius
- Improve its visual effect of showup
- Get alpha value as float

+++ app/src/main/java/com/osfans/trime/data/theme/mapper/GeneralStyleMapper.kt
@@ -141,7 +141,7 @@ class GeneralStyleMapper(
 
         val textFont = getStringList("text_font")
 
-        val textSize = getInt("text_size")
+        val textSize = getFloat("text_size")
 
         val verticalCorrection = getInt("vertical_correction")
